Rename catanatron_env to catanatron_gym
