+++ Makefile
@@ -13,6 +13,10 @@ clean-pyc:
 	find . -name '*.pyc' -exec rm -f {} +
 	find . -name '*.pyo' -exec rm -f {} +
 	find . -name '__pycache__' -exec rm -fr {} +
+	find . -name '*.egg-info' -exec rm -fr {} +
+	find . -name '*.egg' -exec rm -f {} +
+	rm -fr build/
+	rm -fr .pytest_cache/
 	# find . -name '*~' -exec rm -f {} +
 
 clean-test:
