--- catanatron_env/catanatron_env/__init__.py
@@ -1,6 +1,0 @@
-from gymnasium.envs.registration import register
-
-register(
-    id="catanatron_env/Catanatron-v0",
-    entry_point="catanatron_env.envs:CatanatronEnv",
-)
