+++ catanatron_experimental/catanatron_experimental/data_logger.py
@@ -2,11 +2,11 @@ from pathlib import Path
 
 import pandas as pd
 
-from catanatron_env.features import (
+from catanatron_gym.features import (
     create_sample_vector,
     get_feature_ordering,
 )
-from catanatron_env.board_tensor_features import (
+from catanatron_gym.board_tensor_features import (
     CHANNELS,
     HEIGHT,
     WIDTH,
