+++ catanatron_experimental/catanatron_experimental/mcts_score_collector.py
@@ -17,7 +17,7 @@ from catanatron.state_functions import (
 )
 from catanatron_experimental.machine_learning.players.minimax import AlphaBetaPlayer
 from catanatron_experimental.machine_learning.players.playouts import run_playouts
-from catanatron_env.features import (
+from catanatron_gym.features import (
     build_production_features,
     resource_hand_features,
 )
