+++ catanatron_gym/catanatron_gym/__init__.py
@@ -1,0 +1,6 @@
+from gymnasium.envs.registration import register
+
+register(
+    id="catanatron_gym/Catanatron-v0",
+    entry_point="catanatron_gym.envs:CatanatronEnv",
+)
