+++ catanatron_gym/catanatron_gym/envs/__init__.py
@@ -1,0 +1,1 @@
+from catanatron_gym.envs.catanatron_gym import CatanatronEnv
