+++ catanatron_gym/pyproject.toml
@@ -3,10 +3,10 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["catanatron_env"]
+packages = ["catanatron_gym"]
 
 [project]
-name = "catanatron_env"
+name = "catanatron_gym"
 version = "5.0.0"
 description = "Gymnasium for Settlers of Catan"
 readme = "README.md"
