+++ dev-requirements.txt
@@ -1,7 +1,7 @@
 black==25.1.0
 blinker==1.9.0
 -e catanatron_core
--e catanatron_env
+-e catanatron_gym
 -e catanatron_server
 -e catanatron_experimental
 certifi==2025.4.26
