+++ docs/source/sample.rst
@@ -1,0 +1,7 @@
+sample module
+=============
+
+.. automodule:: sample
+   :members:
+   :undoc-members:
+   :show-inheritance:
