+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 testpaths = tests
-pythonpath = catanatron_core catanatron_server catanatron_env catanatron_experimental
+pythonpath = catanatron_core catanatron_server catanatron_gym catanatron_experimental
 addopts = 
     "--benchmark-columns=min,max,mean,stddev,rounds"
     --disable-warnings
