+++ requirements.txt
@@ -1,6 +1,6 @@
 blinker==1.9.0
 -e catanatron_core
--e catanatron_env
+-e catanatron_gym
 -e catanatron_server
 -e catanatron_experimental
 click==8.1.8
