+++ tests/integration_tests/test_speed.py
@@ -4,7 +4,7 @@ from catanatron.game import Game
 from catanatron.json import GameEncoder
 from catanatron.models.player import Color, SimplePlayer, RandomPlayer
 from catanatron.players.weighted_random import WeightedRandomPlayer
-from catanatron_env.features import create_sample
+from catanatron_gym.features import create_sample
 from catanatron_experimental.machine_learning.players.minimax import (
     AlphaBetaPlayer,
     SameTurnAlphaBetaPlayer,
